+++ cmd_windows.go
@@ -12,6 +12,9 @@ func updateCmdByOS(cmd *exec.Cmd) {
 }
 
 func killProcess(p *os.Process) {
+	if p == nil {
+		return
+	}
 	exec.Command("taskkill", "/PID", strconv.Itoa(p.Pid), "/T", "/F").Run()
 	p.Kill()
 }
